+++ apps/publikator/app/authors/page.tsx
@@ -5,8 +5,7 @@ import { useState, useEffect } from 'react'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import {
-  Avatar,
-  Table,
+
   Box,
   Flex,
   Text,
@@ -16,7 +15,6 @@ import {
   TextField,
 } from '@radix-ui/themes'
 import { Search, Plus, ChevronLeft, ChevronRight, X } from 'lucide-react'
-import TooltipIcons from '../components/ui/tooltip-icons'
 import {
   ContributorsDocument,
   ContributorsQuery,
@@ -24,10 +22,10 @@ import {
   OrderDirection,
   ContributorOrderField,
 } from '../../graphql/republik-api/__generated__/gql/graphql'
+import AuthorTable from '../components/author-table'
 
 const AuthorsPage: React.FC = () => {
   const pageSize = 10
-  const router = useRouter()
 
   // Pagination state
   const [currentPage, setCurrentPage] = useState(1)
@@ -167,76 +165,7 @@ const AuthorsPage: React.FC = () => {
         </TextField.Root>
       </Flex>
 
-      <Table.Root>
-        <Table.Header>
-          <Table.Row>
-            <Table.ColumnHeaderCell>Bild</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell>Name</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell>Kurzbio</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell>Info</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify='end'>
-              Aktualisiert
-            </Table.ColumnHeaderCell>
-          </Table.Row>
-        </Table.Header>
-
-        <Table.Body>
-          {contributorsList.map((contributor) => (
-            <Table.Row
-              key={contributor.id}
-              align='center'
-              style={{
-                cursor: 'pointer',
-                transition: 'background-color 0.2s ease',
-              }}
-              onMouseEnter={(e) => {
-                e.currentTarget.style.backgroundColor = '#f8f9fa'
-              }}
-              onMouseLeave={(e) => {
-                e.currentTarget.style.backgroundColor = 'transparent'
-              }}
-              onClick={() => router.push(`/authors/${contributor.slug}`)}
-            >
-              <Table.Cell justify='start'>
-                <Avatar
-                  src={contributor.image || undefined}
-                  fallback={contributor.name?.charAt(0) || '?'}
-                  size='2'
-                />
-              </Table.Cell>
-
-              <Table.Cell justify='start'>
-                <Text weight='medium'>{contributor.name}</Text>
-              </Table.Cell>
-
-              <Table.Cell justify='start'>
-                <Text
-                  style={{
-                    display: '-webkit-box',
-                    WebkitLineClamp: 1,
-                    WebkitBoxOrient: 'vertical',
-                    overflow: 'hidden',
-                    textOverflow: 'ellipsis',
-                    maxWidth: '400px',
-                  }}
-                >
-                  {contributor.shortBio || '—'}
-                </Text>
-              </Table.Cell>
-
-              <Table.Cell justify='center'>
-                <TooltipIcons contributor={contributor} />
-              </Table.Cell>
-
-              <Table.Cell justify='end'>
-                <Text size='1' color='gray' style={{ textAlign: 'right' }}>
-                  {new Date(contributor.updatedAt).toLocaleDateString()}
-                </Text>
-              </Table.Cell>
-            </Table.Row>
-          ))}
-        </Table.Body>
-      </Table.Root>
+      <AuthorTable contributorsList={contributorsList} />
 
       {contributorsList.length === 0 && !loading && (
         <Box p='6' style={{ textAlign: 'center' }}>
@@ -264,7 +193,7 @@ const AuthorsPage: React.FC = () => {
           {debouncedSearchTerm ? (
             <>
               Zeige {startIndex} bis {endIndex} von {totalCount} Suchergebnissen
-              für "{debouncedSearchTerm}"
+              für &quot;{debouncedSearchTerm}&quot;
             </>
           ) : (
             <>
